requirements.txt
atomicwrites==1.4.0
attrs==21.4.0
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
cryptography==37.0.2
docker==5.0.3
Flask==2.1.2
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
idna==3.3
importlib-metadata==4.11.4
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
packaging==21.3
pluggy==1.0.0
py==1.11.0
pycparser==2.21
PyJWT==2.4.0
pyparsing==3.0.9
pytest==7.1.2
#pywin32==304
requests==2.28.0
SQLAlchemy==1.4.37
tomli==2.0.1
urllib3==1.26.9
websocket-client==1.3.2
Werkzeug==2.1.2
zipp==3.8.0
